organize engine/tests/common to submodules
